refactor(results): make results/ outputs-only; move tools to top-level\n\n- Move sweep runner to pipeline/parallel_sweep.py, update README\n- Move parquet loader to analysis/epoch_agg_loader.py\n- Move legacy plotting + helpers to legacy/ (remove shims)\n- Keep only results/parquet and results/sweep under results/\n- Relocate notebooks to top-level notebooks/
