+++ analysis/__init__.py
@@ -1,0 +1,1 @@
+"""Analysis helpers for Parquet outputs and summaries."""
