+++ configs/grid_sweep_1.yaml
@@ -5,17 +5,17 @@ num_trajs: [200]
 num_epochs: [100]
 agents.0.policy: [
   REINFORCE_Gaussian_MarketEnv_Continuous,
-  REINFORCE_Gaussian_v2_MarketEnv_Continuous,
-  REINFORCE_Gaussian_v2_1_MarketEnv_Continuous,
-  REINFORCE_Gaussian_v3_MarketEnv_Continuous,
-  REINFORCE_Gaussian_v3_1_MarketEnv_Continuous,
-  REINFORCE_Gaussian_v4_MarketEnv_Continuous,
-  REINFORCE_Tabu_Gaussian_MarketEnv_Continuous,
-  REINFORCE_Tabu_Gaussian_v2_MarketEnv_Continuous,
-  REINFORCE_Tabu_Gaussian_v3_MarketEnv_Continuous,
-  REINFORCE_Uniform_MarketEnv_Continuous,
-  REINFORCE_Triangular_MarketEnv_Continuous,
-  REINFORCE_Baseline_Gaussian_MarketEnv_Continuous,
+  # REINFORCE_Gaussian_v2_MarketEnv_Continuous,
+  # REINFORCE_Gaussian_v2_1_MarketEnv_Continuous,
+  # REINFORCE_Gaussian_v3_MarketEnv_Continuous,
+  # REINFORCE_Gaussian_v3_1_MarketEnv_Continuous,
+  # REINFORCE_Gaussian_v4_MarketEnv_Continuous,
+  # REINFORCE_Tabu_Gaussian_MarketEnv_Continuous,
+  # REINFORCE_Tabu_Gaussian_v2_MarketEnv_Continuous,
+  # REINFORCE_Tabu_Gaussian_v3_MarketEnv_Continuous,
+  # REINFORCE_Uniform_MarketEnv_Continuous,
+  # REINFORCE_Triangular_MarketEnv_Continuous,
+  # REINFORCE_Baseline_Gaussian_MarketEnv_Continuous,
   # REINFORCE_Baseline_Gaussian_v2_MarketEnv_Continuous,
   # REINFORCE_Baseline_Gaussian_v3_MarketEnv_Continuous,
   # REINFORCE_Baseline_Gaussian_v3_1_MarketEnv_Continuous,
