+++ legacy/__init__.py
@@ -1,0 +1,1 @@
+"""Legacy helpers for CSV/log based analysis used by one_camp_n_days.py."""
