+++ legacy/log_to_csv
@@ -1,17 +1,13 @@
 #!/bin/bash
 
-
-
-## ARGUMENTS ##
-
 log_file="$1"
 output_dir="$2"
 delete_orig="$3"
 
 if [ $# -lt 2 ]; then
-    echo "Not enough arguments provided"
-    echo 'Usage: ' "`basename "$0"`" '<log file> <output dir> <delete original (optional)>'
-    exit 1
+	echo "Not enough arguments provided"
+	echo 'Usage: ' "`basename "$0"`" '<log file> <output dir> <delete original (optional)>'
+	exit 1
 fi
 
 echo '============= START ============='
@@ -22,16 +18,16 @@ policies="`grep -i "agent policies:" "${log_file}" | grep -io "derby.core.polici
 
 if [[ -z "${policies}" ]]
 then
-    if [[ "${experiment}" == "exp_6" ]]
-    then
-        policies="REINFORCE_MarketEnv_Continuous_Gaussian,DummyPolicy1"
-    elif [[ "$experiment" == "exp_7" ]]
-    then
-        policies="REINFORCE_Baseline_MarketEnv_Continuous_Gaussian,DummyPolicy1"
-    else
-        ais="`grep -i "agent policies:" "${log_file}" | sed "s/[[:space:]]//g" | sed "s/),/)---/" | sed "s/,/;/g" | grep -io "AuctionItemSpecification(.*);"`"
-        policies="`grep -i "agent policies:" "${log_file}" | sed "s/[[:space:]]//g" | sed "s/),/)---/" | sed "s/,/;/g" | sed "s/auction_item_spec:${ais}/$(echo "${ais}" | grep -io "uid:[[0-9]]*");/" | sed "s/---/,/g" | sed "s/agentpolicies://" | tr -d []`"
-    fi
+	if [[ "${experiment}" == "exp_6" ]]
+	then
+		policies="REINFORCE_MarketEnv_Continuous_Gaussian,DummyPolicy1"
+	elif [[ "$experiment" == "exp_7" ]]
+	then
+		policies="REINFORCE_Baseline_MarketEnv_Continuous_Gaussian,DummyPolicy1"
+	else
+		ais="`grep -i "agent policies:" "${log_file}" | sed "s/[[:space:]]//g" | sed "s/),/)---/" | sed "s/,/;/g" | grep -io "AuctionItemSpecification(.*);"`"
+		policies="`grep -i "agent policies:" "${log_file}" | sed "s/[[:space:]]//g" | sed "s/),/)---/" | sed "s/,/;/g" | sed "s/auction_item_spec:${ais}/$(echo "${ais}" | grep -io "uid:[[0-9]]*");/" | sed "s/---/,/g" | sed "s/agentpolicies://" | tr -d []`"
+	fi
 fi
 
 # Rename the policies to be clear and concise
@@ -54,7 +50,7 @@ params="`grep "days per traj:" "${log_file}" | sed "s/days per traj://g" | sed "
 lr="`grep -o "learning_rate:.*" "${log_file}" | sed "s/learning_rate://g" | tr -d " "`"
 if [[ -z "${lr}" ]]
 then
-     lr="`grep -io "lr: [0-9a-zA-Z-]*" ${log_file} | sed "s/lr://" | tr -d " "`"
+	 lr="`grep -io "lr: [0-9a-zA-Z-]*" ${log_file} | sed "s/lr://" | tr -d " "`"
 fi
 
 TIMESTAMP=`date "+%Y%m%d%H%M%S"`
@@ -65,7 +61,7 @@ echo "${policies}" > ${csv_file}
 grep "epoch:" ${log_file} | grep -o "\[.*\]" | tr -d "[]()'" | sed "s/agent[0-9]*,//g" | tr -d " " >> ${csv_file}
 if [[ -n "`grep "Traceback (most recent call last)" "${log_file}"`" ]]
 then
-    echo "FAILED" >> "${csv_file}"
+	echo "FAILED" >> "${csv_file}"
 fi
 
 copy_file="${new_file_name}.log"
@@ -76,7 +72,7 @@ mv "${csv_file}" "${output_dir}"
 
 if [[ -n "${delete_orig}" ]]
 then
-    rm -f "${log_file}"
+	rm -f "${log_file}"
 fi
 
 echo "done: ${log_file} => ${output_dir}/${copy_file}, ${output_dir}/${csv_file}"
