+++ legacy/logs_to_csvs
@@ -1,0 +1,18 @@
+#!/bin/bash
+
+log_dir="$1"
+delete_orig="$2"
+
+if [ $# -lt 1 ]; then
+	echo "Not enough arguments provided"
+	echo 'Usage: ' "`basename "$0"`" '<log dir> <delete orig? (optional)>'
+	exit 1
+fi
+
+echo '============= START ============='
+log_dir="$(echo "${log_dir}" | sed 's:/*$::')"
+run_dir="$(basename ${log_dir})"
+results_dir="${log_dir%/}/converted"
+mkdir -p "${results_dir}"
+find "${log_dir}" -type f -iname "*.log" | xargs -I {} "$(dirname "$0")/log_to_csv" {} "${results_dir}" "${delete_orig}"
+echo '============ = END = ============'
