+++ notebooks/README.md
@@ -1,0 +1,8 @@
+# Notebooks
+
+Interactive notebooks for plotting and exploration.
+
+- Plot Results.ipynb
+- Policy Sweep Plot.ipynb
+
+Note: prefer using the Parquet pipeline and `results/analysis` utilities where possible.
