--- results/csvs_to_plots
@@ -1,20 +1,0 @@
-#!/bin/bash
-
-
-
-## ARGUMENTS ##
-
-input_dir="$1"
-output_dir="$2"
-plot_var="$3"
-
-if [ $# -lt 3 ]; then
-    echo "Not enough arguments provided"
-    echo 'Usage: ' "`basename "$0"`" '<input dir> <output dir> <plot var>'
-    exit 1
-fi
-
-echo "plotting method_1"
-python plot_results.py method_1 "${input_dir}" "${output_dir}" "${plot_var}"
-echo "plotting method_2"
-python plot_results.py method_2 "${input_dir}" "${output_dir}" "${plot_var}"
